Python/server/requirements.txt
confluent_kafka==2.3.0
Flask==3.0.0
python-telegram-bot==20.7
Requests==2.31.0
